--- internal/ui/views/createplaylists/createplaylists.go
@@ -1,1 +1,0 @@
-package createplaylists
