--- internal/ui/views/editplaylists/editplaylists.go
@@ -1,1 +1,0 @@
-package editplaylists
