+++ internal/ui/views/editview/editview.go
@@ -1,0 +1,1 @@
+package editview
